BDA696 Final
This folder includes file that run Python nd SQL script and save the output in "results" folder. For some reason, when the docker compose up command runs, it will have an error message "Unknown column FinalScore" but while the you wait a few seconds the SQL script will be read and everything should be running smoothly.